Add log rotation to all containers
Log rotation is a good practice to prevent `no space left on device`.
This can be an attack vector on relayer with malformed requests which can consume all the disk space.

Read more on log rotation in docker-compose:
https://docs.docker.com/compose/compose-file/#logging

```
The default driver json-file, has options to limit the amount of logs stored. To do this, use a key-value pair for maximum storage size and maximum number of files:

options:
  max-size: "200k"
  max-file: "10"
The example shown above would store log files until they reach a max-size of 200kB, and then rotate them. The amount of individual log files stored is specified by the max-file value. As logs grow beyond the max limits, older log files are removed to allow storage of new logs.
```